Refactor: extract metrics query editor and functions editor
